# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Dynamic PDB API
  version: 1.0.0
  description: |
    Dynamic PDB exposes deposited structural biology data, derived models, the
    programs that produced them, and evaluation metrics through a versioned
    JSON:API interface.

    Published entries can be read without authentication. Mutating operations —
    including deposits, model submissions, and review decisions — require a
    bearer token obtained through GitHub sign-in.

    Every request and response is a JSON:API document: send and accept
    `application/vnd.api+json`.
servers:
  - url: https://dynamicpdb.com/api
    description: Production
...
paths:
...
  /v1/auth/github/exchange:
    post:
      tags: [Authentication]
      operationId: exchangeGithubToken
      summary: Exchange GitHub token
      requestBody:
        required: true
        content:
          application/vnd.api+json:
            schema:
              $ref: "#/components/schemas/GitHubExchangeRequest"
      responses:
        "200":
          description: JWT issued.
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/TokenDocument"
        "400":
          description: Invalid request.
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: GitHub access token is invalid or expired.
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Access denied.
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/auth/github/code/exchange:
    post:
      tags: [Authentication]
      operationId: exchangeGithubCode
      summary: Exchange GitHub code
      requestBody:
        required: true
        content:
          application/vnd.api+json:
            schema:
              $ref: "#/components/schemas/GitHubCodeExchangeRequest"
      responses:
        "200":
          description: JWT issued.
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/TokenDocument"
        "400":
          description: Invalid request.
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: GitHub authorization code is invalid or expired.
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Access denied.
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [errors]
      properties:
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ErrorObject"
    ErrorObject:
      type: object
      required: [status]
      properties:
        status:
          type: string
        code:
          type: string
        title:
          type: string
        detail:
          type: string
...
    TokenDocument:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/TokenData"
    TokenData:
      type: object
      required: [type, attributes]
      properties:
        type:
          type: string
        attributes:
          $ref: "#/components/schemas/TokenAttributes"
...
    GitHubExchangeRequest:
      type: object
      required: [access_token]
      properties:
        access_token:
          type: string
    GitHubCodeExchangeRequest:
      type: object
      required: [code, redirect_uri]
      properties:
        code:
          type: string
        redirect_uri:
          type: string
    TokenAttributes:
      type: object
      required: [token_type, access_token, expires_at, name, email, login, permissions]
      properties:
        token_type:
          type: string
          enum: [Bearer]
        access_token:
          type: string
        expires_at:
          type: string
          format: date-time
        name:
          type: string
        email:
          type: string
        login:
          type: string
        permissions:
          type: array
          items:
            type: string
...